src/main/java/com/sting/RepeatedString.java
package com.sting;

import java.io.*;
import java.math.*;
import java.security.*;
import java.text.*;
import java.util.*;
import java.util.concurrent.*;
import java.util.function.*;
import java.util.regex.*;
import java.util.stream.*;
import static java.util.stream.Collectors.joining;
import static java.util.stream.Collectors.toList;

public class RepeatedString {

    static long repeatedString(String s, long n) {

        long count = s.chars().filter(ch -> ch == 'a').count();
        int length = s.length();
        long repeat = (n/length)*count;
        int remain = (int) (n%length);
        if(remain>0)
            repeat += s.substring(0,remain).chars().filter(ch -> ch == 'a').count();
        return repeat;
    }

    public static void main(String[] args) throws IOException {
        BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
        BufferedWriter bufferedWriter = new BufferedWriter(new OutputStreamWriter(System.out));

        String s = bufferedReader.readLine();

        long n = Long.parseLong(bufferedReader.readLine().trim());

        long result = repeatedString(s, n);

        bufferedWriter.write(String.valueOf(result));
        bufferedWriter.newLine();

        bufferedReader.close();
        bufferedWriter.close();
    }
}
